% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_nomenclature.R
\name{apply_nomenclature}
\alias{apply_nomenclature}
\title{Applying nomenclature splits to placement names}
\usage{
apply_nomenclature(tbl)
}
\arguments{
\item{tbl}{Dataframe that includes a column with the placement name}
}
\value{
Dataframe
}
\description{
Applying nomenclature splits to placement names
}
\examples{
\dontrun{
sprinklr_data |> apply_nomenclature()
}
}
