% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decay_build.R
\name{calculate_rmse}
\alias{calculate_rmse}
\title{This function is called within the \verb{fit_decay models()} function; see below}
\usage{
calculate_rmse(actual, predicted)
}
\arguments{
\item{actual}{Actual values of the DV}

\item{predicted}{Predicted values of the DV based on regression}
}
\value{
RMSE value
}
\description{
This function is called within the \verb{fit_decay models()} function; see below
}
\examples{
\dontrun{
calculate_rmse(data[[dv]], exp_predicted)
}
}
