% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_headers.R
\name{clean_headers}
\alias{clean_headers}
\title{Clean the column names of Sprinklr data}
\usage{
clean_headers(tbl)
}
\arguments{
\item{tbl}{Dataframe with Sprinklr names}
}
\value{
Dataframe with updated names
}
\description{
Sprinklr has idiosyncratic column names that depend on the platform
}
\examples{
\dontrun{
sprinklr_data |> clean_headers()
}
}
