% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decay_build.R
\name{decay_build}
\alias{decay_build}
\title{Estimate decay rates for various DVs on a social platform}
\usage{
decay_build(data, group_vars, dv)
}
\arguments{
\item{data}{A merged dataset for a platform, normally built through \code{merge_all_lobs()} or \code{xfi_merge()}}

\item{group_vars}{Variables to group by, e.g. for TikTok c("platform", "ad_objective", "creative_detail", "date")}

\item{dv}{What is the DV that we're estimating the decay}
}
\value{
named list - plot_list is the daily data to be used for plotting; results table is the dataframe with results
}
\description{
Given platform data and a DV of interest, this will estimate the exponential and logarithmic decay of the DV for each piece of creative.
Table results provide the decay rate, half-life (only interpretable as such for exponential), whether regression is significant, RMSE,
r-squared, and AIC. An ID is also provided for easy plotting - see ?fit_decay_model to see how to plot
}
\examples{
\dontrun{
# tiktok
decay_tiktok <- decay_build(merged_tiktok,
group_vars = c("platform", "ad_objective", "creative_detail", "date"),
dv = "avg_duration")

# meta
decay_meta <- decay_build(subset(merged_meta, ad_media_type == "Video"),
group_vars = c("platform", "platform_position", "ad_objective", "creative_detail", "date"),
dv = "avg_duration")
}
}
