% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decay_build.R
\name{fit_decay_models}
\alias{fit_decay_models}
\title{Fit multiple decay models}
\usage{
fit_decay_models(data, dv, day, plot = FALSE)
}
\arguments{
\item{data}{Platform data of a single creative asset summarised to daily level along with other grouping variables}

\item{dv}{DV of interest}

\item{day}{Variable containing numeric count of days that creative ran}

\item{plot}{Logical, TRUE for plot}
}
\value{
Tibble of results, one row is exponential model results, other is logarithmic
}
\description{
This is the main function to fit models and return results. If plot = TRUE, it will retrun a ggplot with the predicted vs actuals
}
\examples{
\dontrun{
fit_decay_models(data = decay_meta$plot_list[[22]], dv = "avg_duration", day = "day", plot = TRUE)
}
}
