% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matrix_data.R
\name{get_matrix_data}
\alias{get_matrix_data}
\title{Read in the line items from Social Matrix sheets}
\usage{
get_matrix_data(
  master_ids,
  line = c("Central", "Internet/Product Diff", "NED", "Retargeting", "West",
    "Xfinity Mobile"),
  platform = c("LinkedIn", "Meta", "Nextdoor", "Pinterest", "Reddit", "Snapchat",
    "TikTok", "Twitter"),
  email
)
}
\arguments{
\item{master_ids}{The name of the master matrix_ids dataframe in working memory}

\item{line}{The division or type of ad account}

\item{platform}{The platform of matrix data}

\item{email}{Your GSP gmail to authenticate and access the sheets}
}
\value{
Data frame with LOB, Platform, link to the Matrix, name of the sheet, and the matrix data
}
\description{
Read in the line items from Social Matrix sheets
}
\examples{
\dontrun{
test <- get_matrix_data(master_ids, line = "West",
platform = "Meta", email = "taylor_grant@gspsf.com")
}
}
