% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSPxfi-package.R
\docType{package}
\name{GSPxfi-package}
\alias{GSPxfi}
\alias{GSPxfi-package}
\title{GSPxfi: Data processing for paid media}
\description{
Merge and summarize paid media data from Sprinklr and Google Sheets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/taylorgrant/GSPxfi}
  \item Report bugs at \url{https://github.com/taylorgrant/GSPxfi/issues}
}

}
\author{
\strong{Maintainer}: Taylor Grant \email{taylor_grant@gspsf.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
