% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massive_cw.R
\name{massive_cw}
\alias{massive_cw}
\title{Massive case_when function}
\usage{
massive_cw(tbl)
}
\arguments{
\item{tbl}{Dataframe of Sprinklr data that includes a column with the placement name}
}
\value{
Dataframe with separated placement names
}
\description{
Placement nomenclatures occasionally fall apart. This gets around that. Clunky but it works
}
\examples{
\dontrun{
sprinklr_data |> massive_cw()
}
}
