% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merged_summarise.R
\name{merged_summarise}
\alias{merged_summarise}
\title{Summarise the merged data}
\usage{
merged_summarise(data, group_vars)
}
\arguments{
\item{data}{Merged dataset}

\item{group_vars}{Variables to group by}
}
\value{
Dataframe with summarized media KPIs
}
\description{
This function allows the user to specify what to group by and then summarises all metrics within the platform. It also calculates overall averages for both video and static for each grouping
}
\examples{
\dontrun{
meta_summary <- dat_summarise(merged_meta,
group_vars = c("platform", "ad_media_type", "platform_position",
"ad_objective", "creative_detail"))
}
}
