% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decay_build.R
\name{plot_decay_models}
\alias{plot_decay_models}
\title{Plot predicted versus actuals for both models}
\usage{
plot_decay_models(data, dv, day, exp_predicted, log_predicted)
}
\arguments{
\item{data}{Platform data of a single creative asset summarised to daily level along with other grouping variables}

\item{dv}{DV of interest}

\item{day}{Variable containing numeric count of days that creative ran}

\item{exp_predicted}{Predicted values of exponential decay model}

\item{log_predicted}{Predicted values of logarithmic decay model}
}
\description{
Plot predicted versus actuals for both models
}
\examples{
\dontrun{
p <- plot_decay_models(data, dv, day, exp_predicted = exp_predicted, log_predicted = log_predicted)
print(p)
}
}
