% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sprinklr.R
\name{read_sprinklr}
\alias{read_sprinklr}
\title{Read the downloaded Sprinklr data into memory}
\usage{
read_sprinklr(file_location)
}
\arguments{
\item{file_location}{The location where the Sprinklr file is located}
}
\value{
Dataframe
}
\description{
Read the downloaded Sprinklr data into memory
}
\examples{
\dontrun{
sprinklr_data <- read_sprinklr(here::here("data", "Meta-Q324.xlsx"))
}
}
