% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decay_build.R
\name{theme_custom}
\alias{theme_custom}
\title{Custom ggplot theme}
\usage{
theme_custom(
  base_family = "Open Sans",
  base_size = 11.5,
  plot_title_family = base_family,
  plot_title_size = 16,
  plot_title_face = "plain",
  plot_title_margin = 4,
  subtitle_family = "Open Sans",
  subtitle_size = 12,
  subtitle_face = "plain",
  subtitle_margin = 15,
  strip_text_family = base_family,
  strip_text_size = 12,
  strip_text_face = "plain",
  caption_family = "Open Sans",
  caption_size = 9,
  caption_face = "italic",
  caption_margin = 10,
  axis_title_family = base_family,
  axis_title_size = 13,
  axis_title_face = "plain",
  axis_title_just = "rt",
  plot_margin = ggplot2::margin(10, 10, 10, 10),
  panel_spacing = ggplot2::unit(0.5, "lines"),
  grid_col = "#cccccc",
  grid = TRUE,
  axis_col = "#cccccc",
  axis = FALSE,
  ticks = FALSE
)
}
\arguments{
\item{base_family}{Base family}

\item{base_size}{Base size}

\item{plot_title_family}{Title}

\item{plot_title_size}{Size}

\item{plot_title_face}{Title face}

\item{plot_title_margin}{Title margin}

\item{subtitle_family}{Subtitle family}

\item{subtitle_size}{Subtitle size}

\item{subtitle_face}{Subtitle face}

\item{subtitle_margin}{Subtitle margin}

\item{strip_text_family}{Strip text family}

\item{strip_text_size}{Strip text size}

\item{strip_text_face}{Strip text face}

\item{caption_family}{Caption family}

\item{caption_size}{Caption size}

\item{caption_face}{Caption face}

\item{caption_margin}{Caption margin}

\item{axis_title_family}{Axis title family}

\item{axis_title_size}{Axis title size}

\item{axis_title_face}{Axis title face}

\item{axis_title_just}{Axis title justification}

\item{plot_margin}{Plot margin}

\item{panel_spacing}{Panel spacing}

\item{grid_col}{Grid col}

\item{grid}{Grid}

\item{axis_col}{Axis col}

\item{axis}{Axis}

\item{ticks}{Ticks}
}
\value{
ggplot theme
}
\description{
Custom ggplot theme
}
\examples{
\dontrun{
library(ggplot2)
ggplot(mpg, aes(displ, hwy)) +
geom_point() +
theme_custom()
}
}
